requirements.txt
ffmpeg-python>=0.2.0
PyYAML>=5.3.1
quart>=0.12.0
quart-cors>=0.3.0
pymongo>=3.10.1
cerberus>=1.3.2